+++ app/src/main/java/com/duckduckgo/app/trackerdetection/TrackerDetector.kt
@@ -36,11 +36,6 @@ class TrackerDetector @Inject constructor() {
     }
 
     fun shouldBlock(url: String, documentUrl: String, resourceType: ResourceType): Boolean {
-        for (client: Client in clients) {
-            if (client.matches(url, documentUrl, resourceType)) {
-                return true
-            }
-        }
-        return false
+        return clients.any { it.matches(url, documentUrl, resourceType) }
     }
 }
